update queue tests
